requirements.txt
dataframe_image==0.2.2
docopt==0.6.2
easyocr==1.7.1
hydra-core==1.3.2
imageio==2.33.0
matplotlib==3.7.3
ml_metrics==0.1.4
numpy==1.24.4
omegaconf==2.3.0
opencv_python==4.8.1.78
pandas==2.1.3
Pillow==7.2.0
Pillow==7.0.0
Pillow==10.1.0
python_Levenshtein==0.23.0
scikit_learn==1.3.1
scipy==1.11.4
skimage==0.0
textdistance==4.6.0
tqdm==4.66.1
